Refactor: Relocate ping module
